[Webview] Convert some straggling jint->jlong field for 64 bit.
Also roll frameworks/webview to include https://android-review.googlesource.com/#/c/88751/

BUG=354405

Review URL: https://codereview.chromium.org/205533006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@258852 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: b7e4fe4a7fe5703267e95f02e1c51b1c5b047411

+++ android_webview/browser/gpu_memory_buffer_factory_impl.cc
@@ -19,7 +19,7 @@ AwDrawGLFunctionTable* g_gl_draw_functions = NULL;
 
 class GpuMemoryBufferImpl : public gfx::GpuMemoryBuffer {
  public:
-  GpuMemoryBufferImpl(int buffer_id, gfx::Size size)
+  GpuMemoryBufferImpl(long buffer_id, gfx::Size size)
       : buffer_id_(buffer_id),
         size_(size),
         mapped_(false) {
@@ -91,7 +91,7 @@ gfx::GpuMemoryBuffer* GpuMemoryBufferFactoryImpl::CreateGpuMemoryBuffer(
   // GL_RGBA8_OES.
   CHECK_EQ(static_cast<GLenum>(GL_RGBA8_OES), internalformat);
   CHECK(g_gl_draw_functions);
-  int buffer_id = g_gl_draw_functions->create_graphic_buffer(width, height);
+  long buffer_id = g_gl_draw_functions->create_graphic_buffer(width, height);
   if (!buffer_id)
     return NULL;
